requirements.txt
kivy==2.3.0
pillow>=10.0.0
pypdf>=3.15.0
plyer>=2.1.0
